deps/v8/test/inspector/debugger/wasm-gc-anyref-expected.txt

Test inspection of Wasm anyref objects

Running test: test
Calling instantiate function.
Waiting for wasm script to be parsed.
Got wasm script!
Module instantiated.
Tables populated.
Setting breakpoint
{
    columnNumber : 246
    lineNumber : 0
    scriptId : <scriptId>
}
Paused:
Script wasm://wasm/739f5f0a byte offset 246: Wasm opcode 0x01 (kExprNop)
Scope:
at $main (0:246):
 - scope (wasm-expression-stack):
   stack: 
 - scope (local):
   $anyref_local: Struct ((ref $type0))
   $anyref_local2: Array ((ref $type1))
   $anyref_local_i31: null (anyref)
   $anyref_local_null: null (anyref)
 - scope (module):
   instance: exports: "exported_ref_table" (Table), "exported_func_table" (Table), "fill_tables" (Function), "main" (Function)
   module: Module
   functions: "$my_func": (Function), "$fill_tables": (Function), "$main": (Function)
   globals: "$global0": function $my_func() { [native code] } (funcref)
   tables: 
      $import.any_table: 0: Array(2) (anyref), 1: Struct ((ref $type0)), 2: undefined (anyref)
      $import.func_table: 0: function () { [native code] } (funcref), 1: function $my_func() { [native code] } (funcref), 2: undefined (funcref)
      $exported_ref_table: 0: Struct ((ref $type0)), 1: Array ((ref $type1)), 2: undefined (anyref), 3: undefined (anyref)
      $exported_func_table: 0: function external_fct() { [native code] } (funcref), 1: function $my_func() { [native code] } (funcref), 2: undefined (funcref)
at (anonymous) (0:17):
 - scope (global):
   -- skipped globals

exports.main returned. Test finished.
